vec3 to ivec3 and some mtx

+++ Includes/Chunk.hpp
@@ -73,7 +73,7 @@ class Chunk
 		bool exposed_block( int row, int col, int level );
 		int get_block_type(siv::PerlinNoise perlin, int row, int col, int level, int surface_level,
 			bool poppy, bool dandelion, bool blue_orchid, bool allium, bool cornflower, bool pink_tulip,
-			bool tree_gen, std::vector<glm::vec3> & trees);
+			bool tree_gen, std::vector<glm::ivec3> & trees);
 		void generate_blocks( void );
         void fill_vertex_array( void );
         void setup_array_buffer( void );
